==> notebooks/linkedin_job_analysis/part_0_retrieve_proxies.ipynb <==
import requests
import pandas as pd
from bs4 import BeautifulSoup

# Get the page containing the list of proxies
url = 'https://free-proxy-list.net/'
response = requests.get(url)
soup = BeautifulSoup(response.text, 'html.parser')

# Find the table containing the list of proxies
table = soup.find('table')

# Get the rows of the table
rows = table.tbody.find_all('tr')

# Extract the proxy details from each row
proxies = {'ip' : [], 'port' : [], 'type' : [], 'country' : []}
for row in rows:
    columns = row.find_all('td')
    # print(row)
    proxies['ip'].append(columns[0].text)
    proxies['port'].append(columns[1].text)
    proxies['type'].append(columns[4].text)
    proxies['country'].append(columns[3].text)

# Print the list of proxies
# print(proxies)

df = pd.DataFrame.from_dict(proxies)
df.head()

df.to_csv("/workspaces/codespaces-jupyter/data/linkedin_job_analysis/part_0_proxies.csv", sep=";", index=False)
