==> notebooks/linkedin_job_analysis/part_2_clean_skills_dataset.ipynb <==
import pandas as pd

# Fonctions utiles pour la suite du notebook

def lower_and_split_words(x):
    return str(x).lower().split(" ")

def check_skills(sr, skill):
  pos = sr[sr == skill]
  return pos

# Thanks to ask9 - arbazkhan971 for the dataset provided on kaggle
# https://www.kaggle.com/datasets/arbazkhan971/allskillandnonskill

# On ouvre le fichier et on supprime les lignes incomplètes
filepath = "/workspaces/codespaces-jupyter/data/linkedin_job_analysis/part_2_all_data_skill_and_nonskills.csv"
dataset = pd.read_csv(filepath, sep =',')
dataset.dropna(inplace = True)

# On supprime les compétences de plus d'un mot car notre modèle de sait pas les gérer
dataset['skills'] = dataset['Text'].apply(lower_and_split_words)
dataset['ct_skills'] = dataset['skills'].apply(lambda x : len(x) == 1)
skills = dataset[dataset['ct_skills']]['Text']

# On constate 33249 compétences, mais il existe des doublons que l'on supprime également pour un total de 27312 compétences
skills = skills.drop_duplicates()
print(skills.describe())

# On supprime les compétences 3727 compétences de 3 caractères dont 99% de nous interessent pas et polluent le compte des compétences par la suite
small_skills = skills[skills.apply(lambda x : len(str(x)) < 4)]
small_skills.describe()

skills = skills[skills.apply(lambda x : len(str(x)) >= 4)]

# On ajoute manuellement les compétences de 3 caractères ou moins que l'on souhaite garder
# Oublie possible, à ajouter au besoin
skill_to_append = ['db', 'bdd', 'js', 'c++', 'etl', 'elt', 'gcp', 'aws', 'c#', 'f#', 'ann', 'nlp', 'cnn', 'ml', 'vm', 'vdi', 'php', 'asm', 'gcc', 'hat', 'pig', 'sql', 'dwh', 'ci', 'cd', 'ops', 'k8s']
series_to_append = pd.Series(skill_to_append)
skills = skills.append(series_to_append)
skills = skills.rename('skills')

# Total de compétences
skills.describe()

# On sauvegarde notre dataset dans un nouveau fichier pour l'utiliser dans l'extraction des compétences dans les descriptions des offres d'emploi
savepath = "/workspaces/codespaces-jupyter/data/linkedin_job_analysis/part_2_cleaned_skills.csv"
skills.to_csv(savepath, sep=',', index=False)
